Refactor RDF code into its own file
